ml/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
transformers==4.35.2
torch==2.1.1
numpy<2.0.0
python-docx==1.1.0
scikit-learn==1.3.2
aiokafka